requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.11.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
gunicorn==21.2.0
drf-spectacular==0.26.5
boto3==1.34.0
django-storages==1.14.2
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
scikit-learn==1.3.0
textblob==0.17.1
nltk==3.8.1
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dateutil==2.8.2
pytz==2023.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2
responses==0.24.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
